定数化

+++ TOBE！SURAIMU/freeFallParticle.cpp
@@ -52,7 +52,7 @@ void FreeFallParticle::update()
 
 void FreeFallParticle::draw()
 {
-	Sprite::getInstance()->draw( texture_ , position_ , nullptr , alpha_ , 0.0F , Vector2( 1.0F , 1.0F ) , rotate_ , Vector2( kTextureSize / 2.0F , kTextureSize / 2.0F ) );
+	Sprite::getInstance()->draw( texture_ , position_ , nullptr , alpha_ , 0.0F , Vector2( 1.0F , 1.0F ) , static_cast< float >( rotate_ ) , Vector2( kTextureSize / 2.0F , kTextureSize / 2.0F ) );
 }
 
 bool FreeFallParticle::isAlive()
